requirements.txt
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.7
python-dateutil>=2.8.2
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0
loguru>=0.6.0
sqlalchemy>=1.4.0
pyarrow>=6.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0